// assets/arc-0031/server/types.ts
export interface AuthMessage {
  /** The domain name of the Verifier */
  domain: string
  /** Algorand account to authenticate with */
  authAcc: string
  /** Unique random nonce generated by the Verifier */
  nonce: string
  /** Optional, description of the Verifier */
  desc?: string
  /** Optional, metadata */
  meta?: string
}